requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
ezdxf==1.3.4
pydantic-settings>=2
python-dotenv==1.0.1
python-multipart==0.0.9
openai>=1.40.0
matplotlib>=3.8.0
reportlab>=4.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0

python-docx>=1.1.0

pytesseract>=0.3.10
opencv-python>=4.9.0
Pillow>=10.3.0
openpyxl>=3.1.2
ezdxf
fastapi
httpx
matplotlib
openai
opencv-python
openpyxl
Pillow
pydantic
pytesseract
pytest
pytest-asyncio
python-docx
python-dotenv
python-multipart
reportlab
uvicorn
ezdxf
fastapi
httpx
matplotlib
openai
opencv-python
openpyxl
Pillow
pydantic
pytesseract
pytest
pytest-asyncio
python-docx
python-dotenv
python-multipart
reportlab
uvicorn
ezdxf==1.3.4
fastapi==0.115.0
httpx>=0.27.0
matplotlib>=3.8.0
openai>=1.40.0
opencv-python>=4.9.0
openpyxl>=3.1.2
Pillow>=10.3.0
pydantic==2.9.2
pytesseract>=0.3.10
pytest>=8.0.0
pytest-asyncio>=0.23.0
python-docx>=1.1.0
python-dotenv==1.0.1
python-multipart==0.0.9
reportlab>=4.0.0
uvicorn==0.30.6
xlrd
xlrd
psycopg2-binary
sqlalchemy
ezdxf
fastapi
httpx
matplotlib
openai
opencv-python
openpyxl
Pillow
psycopg2-binary
pydantic
pytesseract
pytest
pytest-asyncio
python-docx
python-dotenv
python-multipart
reportlab
sqlalchemy
uvicorn
xlrd
